Revert "deactivate arc interpolation"
This reverts commit bde624b20070aa907b0485dd6f78f4cfec38b0db.
