+++ io/plugins/eu.esdihumboldt.hale.io.gml.geometry/src/eu/esdihumboldt/hale/io/gml/geometry/handler/ArcByCenterPointHandler.java
@@ -13,7 +13,7 @@
  *     wetransform GmbH <http://www.wetransform.to>
  */
 
-package eu.esdihumboldt.hale.io.gml.geometry._disabled_handler_;
+package eu.esdihumboldt.hale.io.gml.geometry.handler;
 
 import java.text.ParseException;
 import java.util.ArrayList;
@@ -45,7 +45,6 @@ import eu.esdihumboldt.hale.io.gml.geometry.GMLGeometryUtil;
 import eu.esdihumboldt.hale.io.gml.geometry.GeometryNotSupportedException;
 import eu.esdihumboldt.hale.io.gml.geometry.InterpolationSupportedGeometryHandler;
 import eu.esdihumboldt.hale.io.gml.geometry.constraint.GeometryFactory;
-import eu.esdihumboldt.hale.io.gml.geometry.handler.PointHandler;
 import eu.esdihumboldt.util.geometry.interpolation.ArcByCenterPointInterpolation;
 import eu.esdihumboldt.util.geometry.interpolation.Interpolation;
