+++ io/plugins/eu.esdihumboldt.hale.io.gml.geometry/src/eu/esdihumboldt/hale/io/gml/geometry/handler/CircleHandler.java
@@ -13,7 +13,7 @@
  *     wetransform GmbH <http://www.wetransform.to>
  */
 
-package eu.esdihumboldt.hale.io.gml.geometry._disabled_handler_;
+package eu.esdihumboldt.hale.io.gml.geometry.handler;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -35,7 +35,6 @@ import eu.esdihumboldt.hale.common.schema.model.constraint.type.Binding;
 import eu.esdihumboldt.hale.common.schema.model.constraint.type.GeometryType;
 import eu.esdihumboldt.hale.io.gml.geometry.GeometryNotSupportedException;
 import eu.esdihumboldt.hale.io.gml.geometry.constraint.GeometryFactory;
-import eu.esdihumboldt.hale.io.gml.geometry.handler.LineStringHandler;
 import eu.esdihumboldt.util.geometry.interpolation.CircleInterpolation;
 import eu.esdihumboldt.util.geometry.interpolation.Interpolation;
